requirements.txt
fastapi
uvicorn
sqlmodel
asyncpg
pydantic
python-dotenv
openai
httpx
youtube-transcript-api
# upstash-qstash
pyjwt
pytest
pytest-asyncio
httpx
pytest-cov
pytest-mock
aiosqlite
notion-client
